requirements.txt
numpy==1.26.4
typeguard 
pexpect
gtts
openai-whisper
faster-whisper
ffmpeg-python
jiwer
requests
werkzeug==2.2.3
sounddevice
python-dotenv
google-generativeai
SpeechRecognition
pyaudio
pydub==0.25.1
unidic
cython>=0.29.30
scipy>=1.11.2
torch==2.7.0
torchvision==0.22.0
torchaudio
soundfile>=0.12.0
librosa>=0.10.0
numba>=0.57.0
tqdm>=4.64.1
flask
deepspeed
vinorm==2.0.7
cutlet
underthesea
gradio==4.35
deepfilternet==0.5.6
